+++ CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.0] - 2025-10-18
 
 ### Breaking Changes
 
+- Change `cmd()` to accept strings or a list of strings for commands.
 - Change `Cmd.run()` to not use `process.stdin` by default as input.
 - Remove `Cmd.runDebug()` since it is now identical to `Cmd.run()`.